requirements.txt
flask-wtf
flask-sqlalchemy
flask-migrate
flask-mail
flask-login
flask-bootstrap
flask-uploads
flask-mysql
flask-mysqldb
numpy
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0